+++ compiler/ext/ext_path.ml
@@ -28,11 +28,6 @@ type t =
   | Dir of string
 [@@unboxed]
 
-let simple_convert_node_path_to_os_path =
-  if Sys.unix then fun x -> x
-  else if Sys.win32 || Sys.cygwin then Ext_string.replace_slash_backward
-  else failwith ("Unknown OS : " ^ Sys.os_type)
-
 let cwd = lazy (Sys.getcwd ())
 
 let split_by_sep_per_os : string -> string list =
@@ -80,61 +75,10 @@ let node_rebase_file ~from ~to_ file =
        else node_relative_path ~from:(Dir from) (Dir to_))
     file
 
-(***
-   {[
-     Filename.concat "." "";;
-     "./"
-   ]}
-*)
-let combine path1 path2 =
-  if Filename.is_relative path2 then
-    if Ext_string.is_empty path2 then path1
-    else if path1 = Filename.current_dir_name then path2
-    else if path2 = Filename.current_dir_name then path1
-    else Filename.concat path1 path2
-  else path2
-
 let ( // ) x y =
   if x = Filename.current_dir_name then y
   else if y = Filename.current_dir_name then x
   else Filename.concat x y
-
-(**
-   {[
-     split_aux "//ghosg//ghsogh/";;
-     - : string * string list = ("/", ["ghosg"; "ghsogh"])
-   ]}
-   Note that
-   {[
-     Filename.dirname "/a/" = "/"
-       Filename.dirname "/a/b/" = Filename.dirname "/a/b" = "/a"
-   ]}
-   Special case:
-   {[
-     basename "//" = "/"
-       basename "///"  = "/"
-   ]}
-   {[
-     basename "" =  "."
-       basename "" = "."
-       dirname "" = "."
-       dirname "" =  "."
-   ]}
-*)
-let split_aux p =
-  let rec go p acc =
-    let dir = Filename.dirname p in
-    if dir = p then (dir, acc)
-    else
-      let new_path = Filename.basename p in
-      if Ext_string.equal new_path Filename.dir_sep then go dir acc
-        (* We could do more path simplification here
-           leave to [rel_normalized_absolute_path]
-        *)
-      else go dir (new_path :: acc)
-  in
-
-  go p []
 
 (**
    TODO: optimization
@@ -143,86 +87,6 @@ let split_aux p =
    This function is useed in [es6-global] and
    [amdjs-global] format and tailored for `rollup`
 *)
-let rel_normalized_absolute_path ~from to_ =
-  let root1, paths1 = split_aux from in
-  let root2, paths2 = split_aux to_ in
-  if root1 <> root2 then root2
-  else
-    let rec go xss yss =
-      match (xss, yss) with
-      | x :: xs, y :: ys ->
-        if Ext_string.equal x y then go xs ys
-        else if x = Filename.current_dir_name then go xs yss
-        else if y = Filename.current_dir_name then go xss ys
-        else
-          let start =
-            Ext_list.fold_left xs Ext_string.parent_dir_lit (fun acc _ ->
-                acc // Ext_string.parent_dir_lit)
-          in
-          Ext_list.fold_left yss start (fun acc v -> acc // v)
-      | [], [] -> Ext_string.empty
-      | [], y :: ys -> Ext_list.fold_left ys y (fun acc x -> acc // x)
-      | _ :: xs, [] ->
-        Ext_list.fold_left xs Ext_string.parent_dir_lit (fun acc _ ->
-            acc // Ext_string.parent_dir_lit)
-    in
-    let v = go paths1 paths2 in
-
-    if Ext_string.is_empty v then Literals.node_current
-    else if
-      v = "." || v = ".."
-      || Ext_string.starts_with v "./"
-      || Ext_string.starts_with v "../"
-    then v
-    else "./" ^ v
-
-(*TODO: could be hgighly optimized later
-  {[
-    normalize_absolute_path "/gsho/./..";;
-
-    normalize_absolute_path "/a/b/../c../d/e/f";;
-
-    normalize_absolute_path "/gsho/./..";;
-
-    normalize_absolute_path "/gsho/./../..";;
-
-    normalize_absolute_path "/a/b/c/d";;
-
-    normalize_absolute_path "/a/b/c/d/";;
-
-    normalize_absolute_path "/a/";;
-
-    normalize_absolute_path "/a";;
-  ]}
-*)
-
-(** See tests in {!Ounit_path_tests} *)
-let normalize_absolute_path x =
-  let drop_if_exist xs =
-    match xs with
-    | [] -> []
-    | _ :: xs -> xs
-  in
-  let rec normalize_list acc paths =
-    match paths with
-    | [] -> acc
-    | x :: xs ->
-      if Ext_string.equal x Ext_string.current_dir_lit then
-        normalize_list acc xs
-      else if Ext_string.equal x Ext_string.parent_dir_lit then
-        normalize_list (drop_if_exist acc) xs
-      else normalize_list (x :: acc) xs
-  in
-  let root, paths = split_aux x in
-  let rev_paths = normalize_list [] paths in
-  let rec go acc rev_paths =
-    match rev_paths with
-    | [] -> Filename.concat root acc
-    | last :: rest -> go (Filename.concat last acc) rest
-  in
-  match rev_paths with
-  | [] -> root
-  | last :: rest -> go last rest
 
 let absolute_path cwd s =
   let process s =
@@ -245,13 +109,6 @@ let absolute_cwd_path s = absolute_path cwd s
    match s with
    | File x -> File (absolute_path cwd x )
    | Dir x -> Dir (absolute_path cwd x) *)
-
-let concat dirname filename =
-  if filename = Filename.current_dir_name then dirname
-  else if dirname = Filename.current_dir_name then filename
-  else Filename.concat dirname filename
-
-let check_suffix_case = Ext_string.ends_with
 
 (* Input must be absolute directory *)
 let rec find_root_filename ~cwd filenames =