+++ compiler/ext/ext_path.mli
@@ -24,56 +24,10 @@
 
 type t
 
-val simple_convert_node_path_to_os_path : string -> string
-(** Js_output is node style, which means
-    separator is only '/'
-
-    if the path contains 'node_modules',
-    [node_relative_path] will discard its prefix and
-    just treat it as a library instead
-*)
-
-val combine : string -> string -> string
-(**
-   [combine path1 path2]
-   1. add some simplifications when concatenating
-   2. when [path2] is absolute, return [path2]
-*)
-
-(**
-   {[
-     get_extension "a.txt" = ".txt"
-       get_extension "a" = ""
-   ]}
-*)
-
 val node_rebase_file : from:string -> to_:string -> string -> string
-
-val rel_normalized_absolute_path : from:string -> string -> string
-(**
-   TODO: could be highly optimized
-   if [from] and [to] resolve to the same path, a zero-length string is returned
-   Given that two paths are directory
-
-   A typical use case is
-   {[
-     Filename.concat
-       (rel_normalized_absolute_path cwd (Filename.dirname a))
-       (Filename.basename a)
-   ]}
-*)
-
-val normalize_absolute_path : string -> string
+(* val node_rebase_file : from:string -> to_:string -> string -> string *)
 
 val absolute_cwd_path : string -> string
-
-val concat : string -> string -> string
-(** [concat dirname filename]
-    The same as {!Filename.concat} except a tiny optimization
-    for current directory simplification
-*)
-
-val check_suffix_case : string -> string -> bool
 
 (* It is lazy so that it will not hit errors when in script mode *)
 val package_dir : string Lazy.t